+++ tests/unit-global/esnext.number.clamp.js
@@ -10,7 +10,6 @@ QUnit.test('Number#clamp', assert => {
   assert.same(clamp.call(4, 2, 6), 4);
   assert.same(clamp.call(6, 2, 4), 4);
 
-  assert.same(clamp.call(NaN, 4, 6), NaN, 'If value is NaN, return NaN.');
   assert.same(clamp.call(-0, 0, 1), 0, 'If value is -0𝔽 and min is +0𝔽, return +0𝔽.');
   assert.same(clamp.call(0, -0, 1), 0, 'If value is +0𝔽 and min is -0𝔽, return +0𝔽.');
   assert.same(clamp.call(-0, -1, 0), -0, 'If value is -0𝔽 and max is +0𝔽, return -0𝔽.');
@@ -20,8 +19,11 @@ QUnit.test('Number#clamp', assert => {
   assert.same(clamp.call(2, 0, -0), -0, 'min is +0𝔽 and max is -0𝔽');
   assert.same(clamp.call(2, 3, 1), 1, 'min > max');
 
+  assert.same(clamp.call(NaN, 3, 1), NaN, 'If value is NaN, return NaN.');
+  assert.same(clamp.call(2, NaN, 1), NaN, 'If min is NaN, return NaN.');
+  assert.same(clamp.call(2, 3, NaN), NaN, 'If max is NaN, return NaN.');
+
+  assert.throws(() => clamp.call({ valueOf: () => 2 }, 1, 3), TypeError, 'If value is not a Number, throw a TypeError exception');
   assert.throws(() => clamp.call(2, Object(1), 3), TypeError, 'If min is not a Number, throw a TypeError exception.');
-  assert.throws(() => clamp.call(2, NaN, 3), RangeError, 'If min is NaN, throw a RangeError exception.');
   assert.throws(() => clamp.call(2, 1, Object(3)), TypeError, 'If max is not a Number, throw a TypeError exception.');
-  assert.throws(() => clamp.call(2, 1, NaN), RangeError, 'If max is NaN, throw a RangeError exception.');
 });