+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 ##### Unreleased
+- `clamp` no longer throws an error on `NaN` as `min` or `max`, following [proposal-math-clamp#d2387791c265edf66fbe2455eab919016717ce6f](https://github.com/tc39/proposal-math-clamp/commit/d2387791c265edf66fbe2455eab919016717ce6f) 
 - Use `Get` in `Iterator.zipKeyed`, following [proposal-joint-iteration#43](https://github.com/tc39/proposal-joint-iteration/pull/43)
 - Fixed [several V8 bugs](https://github.com/zloirock/core-js/issues/1439) in `Uint8Array.fromHex` and `Uint8Array.prototype.{ setFromBase64, toBase64, toHex }`, thanks [**@brc-dd**](https://github.com/brc-dd)
 - Fixed some cases of `Set.prototype.{ symmetricDifference, union }` detection